refactor(useOrder): separate logic with custom hook
